Assignment-1/requirements.txt
flask
pandas
mysql-connector-python
